backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
pydantic-settings
oracledb
pandas
openpyxl
weasyprint
fpdf2
python-multipart
pytest
pytest-mock
pytest-asyncio
passlib[bcrypt]
bcrypt==3.2.2
PyJWT
python-jose[cryptography]
httpx
aiosqlite
